Secureboot: Support signing kernel modules with an ephemeral key
The current sonic secureboot implementation assumes all assets are
signed with the DB Key.  There are roughly 3600 kernel modules that
get signed in the process.

Organizations may have varying security policy requirements where
any signing requests may require signoff by one or more parties,
this may simply be infeasible for that many requests.

Given that the DB key may be a long-lived key, and only RSA 2048
is often used or available, it is a security best-practice to sign
as few assets as possible with such a key.