Switch to using Boost 1.83 for Bookworm
With the base image upgrade to Trixie, Bookworm-based containers will
need to use Boost 1.83. This is because of an incompatibility between
rsyslog_plugin that uses Boost 1.83 on Trixie and the eventd
container that uses Boost 1.74 on Bookworm; specifically there is an
incompatiblity with serialization of objects between the two versions of
Boost.

Because of this, for Bookworm, use Boost 1.83 instead of the default to
the default Boost 1.74.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>
Signed-off-by: xiaweijiang <xiaweijiang@microsoft.com>