Added more clarity about the error message when a reboot is in progress (#22576)
Added more clarity about the error message when a reboot is in progress

Why I did it
Added appropriate reasons for blocking the reboot.

How I did it
Added detailed error message.

How to verify it
Added Testcases to verify the changes.

Description for the changelog
Improved the failure error messages.