[libnss-radius] handle empty secondary groups for useradd
Why I did it
User authentication through radius fails with invalid password information sent to radius server.

How I did it
When a user login is attempted, the user is created locally and then the user is processed for authentication. With the issue, the user creation fails and the successive authentication happens with wrong information. When the user was created locally and then the authentication was attempted, we dont hit the issue.

the user addition fails as the secondary group list is empty and when it is, useradd fails with invalid -G argument. Check for secondary group string validity before using the same.

How to verify it
Verified the user authentication is successful.