Don't attempt building the user container if the main slave fails
If building the main slave container fails, then don't continue and try
to build the user slave container (or try to collect the versions used
in the failed slave container). This reduces the number of errors logged
after the actual error message, and makes it a bit easier for reporting
the actual cause of failures, besides saying that the user slave
container failed to build because the main slave container wasn't
available.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>