Moby512 HWSKU cleanup and new FANOUT HWSKU
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it

As per the request in https://github.com/aristanetworks/sonic-qual.msft/issues/683, adding new HWSKU `Arista-7060X6-16PE-384C-B-O128S2-FANOUT` for Moby512.

This PR also contains couple cleanups in the existing Moby512 HWSKUs
- `Arista-7060X6-16PE-384C-B-O128S2-COPPER-LAB` is removed as it's the same as `Arista-7060X6-16PE-384C-B-O128S2-LAB` and additional dynamic autoneg configs
- Fix the symlinks of `Arista-7060X6-16PE-384C-B-O128S2-LAB` to fix the current orchagent crash when loading this HWSKU

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

Use `Arista-7060X6-16PE-384C-B-O128S2` as a baseline for Moby512 configuration and make corresponding lanemap and polarity changes for `-LAB` and `-FANOUT`

#### How to verify it

Tested internaly on our Moby512 setup
- FANOUT: test with two Moby512 switches setup with one loaded `Arista-7060X6-16PE-384C-B-O128S2` and another loaded `Arista-7060X6-16PE-384C-B-O128S2-FANOUT`, and confirmed backplane ports are up
- LAB: test with a single Moby512 with snake setup between top and bottom backplane ports and confirmed they are up

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [x] 202505
- [x] msft-202412
- [x] msft-202503

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [x] msft-202412

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)