[Mellanox] Add common sai.profile per asic and asic_detect.sh script (#23458)
- Why I did it
This PR adds new script for Mellanox asic detection, as there are few places that use it.
The script queries "lspci" command to find out what is the current device ID.
If the script was successful, we write the value to asic_type so the script will be running only 1 time.

Also, added sai-spcx.profile and sai-bf3.profile with relevant parameters.

- How I did it
Added new script to detect asic type, and new files for per-asic sai.profile

- How to verify it
Run on switch and make sure the parameters are applied to sai.profile inside syncd.

Signed-off-by: Feng Pan <fenpan@microsoft.com>