Fix subport of Moby backplane ports (#23600)
Why I did it
As per the issue reported in aristanetworks/sonic-qual.msft#594, Moby backplane ports always have the subport to be 0, but we would like them to be indexed from 0 to 12. This PR fixes various issue with existing subport computation so the correct subport can be used for these Moby backplane ports.

How I did it
Update portconfig script to make sure it can use the values defined in hwsku.json when they are available, and compute the subports based on the new port naming so that the values can also be computed correctly in the future for the products like Moby instead of always giving 0 by default
Update hwsku.json with the new subport for Moby backplane ports. Note that based on Moby512 HWSKU cleanup and new FANOUT HWSKU #23375 and Moby HWSKU cleanup #23599, we only need to update a single hwsku.json file for each platform
Update yang model accordingly so that subport 9-12 can be accepted

How to verify it
Tested on the Moby switch in our lab to make sure the correct subports are created in config file and config reload can also preserve these correct subport configuration.

Co-authored-by: bobbymcgonigle <bobbymcgonigle@arista.com>