[Mellanox][Smartswitch] Add Force power indication in the status for dpuctl
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it
The force power sysfs file relevant for the DPU is readable (to check if force power is required or not) This sysfs file is displayed as part of the
`dpuctl dpu-status` CLI, this helps in debugging for the smartswitch platforms.

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it
By modifying the main.py file for the dpuctl package to include a function call to the `DpuModule` to read the force power on file for sysfs
#### How to verify it
Unit tests
```
tests/test_dpuctl.py::Testdpuctl::test_dpuctl_power_off PASSED                                                                                                                                                                              [ 24%]
tests/test_dpuctl.py::Testdpuctl::test_dpuctl_power_on PASSED                                                                                                                                                                               [ 25%]
tests/test_dpuctl.py::Testdpuctl::test_dpuctl_reset PASSED                                                                                                                                                                                  [ 25%]
tests/test_dpuctl.py::Testdpuctl::test_dpuctl_status PASSED                                                                                                                                                                                 [ 26%]
```
Tested manually on smartswitch platform

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [ ] 202205
- [ ] 202211
- [ ] 202305
- [ ] 202311
- [ ] 202405
- [ ] 202411
- [ ] 202505

#### Tested branch (Please provide the tested image version)
Master branch
<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [ ] <!-- image version 1 -->
- [ ] <!-- image version 2 -->

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)