[build] Fix sonic-utilities-data submodule dep
<!--
     Please make sure you've read and understood our contributing guidelines:
     https://github.com/Azure/SONiC/blob/gh-pages/CONTRIBUTING.md

     ** Make sure all your commits include a signature generated with `git commit -s` **

     If this is a bug fix, make sure your description includes "fixes #xxxx", or
     "closes #xxxx" or "resolves #xxxx"

     Please provide the following information:
-->

#### Why I did it

The `sonic-utilities-data_1.0-1_all.deb` target depends on sources from the `src/sonic-utilities` submodule. However, the cache dependency file does not list this submodule as a dependency. So, when the build cache is used, updates to the `src/sonic-utilities` submodule pointer do not trigger the necessary rebuild of `sonic-utilities-data_1.0-1_all.deb`, leading to a stale build output.

##### Work item tracking
- Microsoft ADO **(number only)**:

#### How I did it

Updated `rules/sonic-utilities-data.dep` to include the proper `SMDEP` rules.

#### How to verify it

1. Perform a build with the dpkg cache enabled
2. Confirm the new output file `target/debs/bookworm/sonic-utilities-data_1.0-1_all.deb.smdep` is created, containing the submodule dependencies (smdeps)
3. Archive the `target/debs/bookworm/sonic-utilities-data_1.0-1_all.deb` output outside the build workspace
4. Clean the build workspace
5. Update the `src/sonic-utilities` submodule pointer to a new commit that contains different source code
6. Perform another build with the dpkg cache enabled
7. Extract the previous and current `sonic-utilities-data_1.0-1_all.deb` outputs. Confirm that the contents of the debian packages are distinct and contain the expected source based on the respective `src/sonic-utilities` submodule pointers

<!--
If PR needs to be backported, then the PR must be tested against the base branch and the earliest backport release branch and provide tested image version on these two branches. For example, if the PR is requested for master, 202211 and 202012, then the requester needs to provide test results on master and 202012.
-->

#### Which release branch to backport (provide reason below if selected)

<!--
- Note we only backport fixes to a release branch, *not* features!
- Please also provide a reason for the backporting below.
- e.g.
- [x] 202006
-->

- [x] 202405
- [x] 202411
- [x] 202505

This is a build bug which silently results in incorrect build outputs when the dpkg cache is used. As such, it should be double committed.

#### Tested branch (Please provide the tested image version)

<!--
- Please provide tested image version
- e.g.
- [x] 20201231.100
-->

- [x] 202405 (6fad19aff0cf0cc2132db1de92c20d27e1daabe9)

#### Description for the changelog
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->

Fix submodule build dependency for sonic-utilities-data

<!--
 Ensure to add label/tag for the feature raised. example - PR#2174 under sonic-utilities repo. where, Generic Config and Update feature has been labelled as GCU.
-->

#### Link to config_db schema for YANG module changes
<!--
Provide a link to config_db schema for the table for which YANG model
is defined
Link should point to correct section on https://github.com/Azure/sonic-buildimage/blob/master/src/sonic-yang-models/doc/Configuration.md
-->

#### A picture of a cute animal (not mandatory but encouraged)