[FRR]: Add fix for the issue reported in https://github.com/sonic-net/sonic-buildimage/issues/23121
This PR fixes an issue with routes installation in FRR caused by the ordering of events.

* First, a kernel route is received.
* Then, a connected route for the same destination arrives.
* The connected route should win over the kernel route. As a result, FRR is expected to remove the kernel route and keep only the connected route.

However, there are cases where, due to the order in which events are processed, FRR ends up with both a connected route and a kernel route for the same destination.

This PR fixes the issue by ensuring that the kernel route is removed from the early route queue before the connected route is installed.

Fixes https://github.com/sonic-net/sonic-buildimage/issues/23121

Signed-off-by: Carmine Scarpitta <cscarpit@cisco.com>