Merged PR 13346397: Replace use of MSSONIC-TOKEN with access token
Replace use of MSSONIC-TOKEN with access token

The MSSONIC-TOKEN will be deprecated due to the limitation of using service account in Azure DevOps. This change updated the template to use access token of service connection to access the internal project under mssonic organization for downloading the pr_test_scripts.yml file.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
This PR implements an authentication update.

#### PR Summary
The PR modifies the test plan pipeline script to replace the usage of a hard-coded token with dynamic access token retrieval via Azure CLI for secure authentication.
- `/.pipelines/elastictest_template/run-test-elastictest-template.yaml`: Replaces the static MSSONIC-TOKEN based curl call with an AzureCLI@2 task that fetches and applies an access token.
- `/.pipelines/elastictest_template/run-test-elastictest-template.yaml`: Includes minor script formatting cleanups.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->